・ファイル構成変更 ・コマンド名変更(tree→rtree) ・コマンドライン引数の処理をclapに変更
